Add test for post_ssh
